improve ImgLibProcessor and associated classes when working with LongType data
git-svn-id: http://code.imagej.net/svn/imagej/trunk@1504 be23795f-a60e-40d4-ba19-055798872e57

+++ imagej/src/main/java/imagej/process/function/SqrtUnaryFunction.java
@@ -21,7 +21,7 @@ public class SqrtUnaryFunction implements UnaryFunction
 			value = Math.sqrt(input);
 	
 		if (this.dataIsIntegral)
-			value = (int) value;
+			value = (long) value;
 		
 		return value;
 	}
